Authorization issue

+++ api/group.py
@@ -2,13 +2,32 @@ from flask import Blueprint, request, jsonify, abort
 from db import get_session
 from models import *
 from schemas import *
-group = Blueprint('group', __name__, url_prefix='/groups')
+from flask_httpauth import HTTPBasicAuth
+from werkzeug.security import check_password_hash
 
+group = Blueprint('group', __name__, url_prefix='/groups')
+auth = HTTPBasicAuth()
+
+@auth.verify_password
+def verify_password(username, password):
+    session = get_session()
+    user = session.query(User).filter(User.email == username).first()
+    if not user:
+        return False
+    if check_password_hash(user.password, password):
+        return True
+    return False
+
+def get_current_user() -> User:
+    Session = get_session()
+    username = auth.username()
+    return Session.query(User).filter(User.email==username).first()
 # temporary using query parameters for user id
 @group.route('/', methods=['POST', 'GET'])
+@auth.login_required
 def groups():
     session = get_session()
-    user_id = request.args.get("user_id") #temporary
+    user_id = get_current_user().id #temporary
 
     if request.method == 'POST':
         try:
@@ -38,6 +57,7 @@ def groups():
         return jsonify(my_groups)
 
 @group.route('/<group_id>', methods=['GET', 'PUT', 'DELETE'])
+@auth.login_required
 def group_by_id(group_id):
     session = get_session()
     if not group_id.isnumeric():
@@ -45,11 +65,25 @@ def group_by_id(group_id):
 
     group = session.query(Group).filter(Group.id==group_id).first()
     if not group:
-        abort(404)
+        return {
+                'code': 'group',
+                'message': 'Group not found',
+            }, 404
 
     if request.method == 'GET':
+        membership = session.query(GroupMember).filter(GroupMember.user_id==get_current_user().id, GroupMember.group_id==group_id).first()
+        if membership is None:
+            return {
+                'code': 'user',
+                'message': 'Unauthorized user access/No rights',
+            }, 401
         return jsonify(GroupInfoSchema().dump(group))
     elif request.method == 'PUT':
+        if get_current_user().id != group.owner_id:
+            return {
+                'code': 'user',
+                'message': 'Unauthorized user access/No rights',
+            }, 401
         updated_info = dict()
         print(request.args.get("name"))
         if request.args.get("name"):
@@ -63,11 +97,17 @@ def group_by_id(group_id):
         session.refresh(group)
         return jsonify(GroupInfoSchema().dump(group))
     elif request.method == 'DELETE':
+        if get_current_user().id != group.owner_id:
+            return {
+                'code': 'user',
+                'message': 'Unauthorized user access/No rights',
+            }, 401
         session.query(Group).filter(Group.id==group.id).delete()
         session.commit()
         return jsonify({'Message': 'There is no such group now!'})
 
 @group.route('/<group_id>/members', methods=['GET', 'POST', 'DELETE'])
+@auth.login_required
 def group_members(group_id):
     session = get_session()
     if not group_id.isnumeric():
@@ -75,7 +115,10 @@ def group_members(group_id):
 
     group = session.query(Group).filter(Group.id==group_id).first()
     if not group:
-        abort(404)
+        return {
+                'code': 'group',
+                'message': 'Group not found',
+            }, 404
 
     def get_groupmember_schema(member, user):
         member_schema = GroupMemberInfoSchema().dump(member)
@@ -85,6 +128,11 @@ def group_members(group_id):
         return member_schema
 
     if request.method == 'GET':
+        if get_current_user().id != group.owner_id:
+            return {
+                'code': 'user',
+                'message': 'Unauthorized user access/No rights',
+            }, 401
         members = session.query(GroupMember).filter(
             GroupMember.group_id==group_id).all()
         res = []
@@ -94,6 +142,11 @@ def group_members(group_id):
         return jsonify(res)
     
     elif request.method == 'POST':
+        if get_current_user().id != group.owner_id:
+            return {
+                'code': 'user',
+                'message': 'Unauthorized user access/No rights',
+            }, 401
         email = request.args.get("email")
         user = session.query(User).filter(User.email==email).first()
         if not user:
@@ -105,12 +158,18 @@ def group_members(group_id):
         session.commit()
         return jsonify(get_groupmember_schema(new_member, user))
     elif request.method == 'DELETE':
+        if get_current_user().id != group.owner_id:
+            return {
+                'code': 'user',
+                'message': 'Unauthorized user access/No rights',
+            }, 401
         member_id = request.args.get("member_id")
         session.query(GroupMember).filter(GroupMember.id==member_id).delete()
         session.commit()
         return jsonify({'Message': 'There is no such group member now!'})
 
 @group.route('/<group_id>/tasks', methods=['GET', 'POST'])
+@auth.login_required
 def group_tasks(group_id):
     session = get_session()
     if not group_id.isnumeric():
@@ -118,15 +177,29 @@ def group_tasks(group_id):
         
     group = session.query(Group).filter(Group.id==group_id).first()
     if not group:
-        abort(404)
+        return {
+                'code': 'group',
+                'message': 'Group not found',
+            }, 404
 
     if request.method == 'GET':
+        membership = session.query(GroupMember).filter(GroupMember.user_id==get_current_user().id, GroupMember.group_id==group_id).first()
+        if membership is None:
+            return {
+                'code': 'user',
+                'message': 'Unauthorized user access/No rights',
+            }, 401
         group_tasks = session.query(GroupTask).filter(GroupTask.group_id==group_id).all()
         tasks = []
         for task in group_tasks:
             tasks.append(GroupTaskInfoSchema().dump(task))
         return jsonify(tasks)
     elif request.method == 'POST':
+        if get_current_user().id != group.owner_id:
+            return {
+                'code': 'user',
+                'message': 'Unauthorized user access/No rights',
+            }, 401
         try:
             new_task = CreateTaskSchema().load(request.json)
             new_task.group_id = group_id
@@ -150,6 +223,7 @@ def group_tasks(group_id):
         return jsonify(GroupTaskInfoSchema().dump(new_task))
 
 @group.route('/<group_id>/tasks/<task_id>', methods=['GET', 'DELETE'])
+@auth.login_required
 def delete_group_task(group_id, task_id):
     session = get_session()
     group = session.query(Group).filter(Group.id==group_id).first()
@@ -158,11 +232,22 @@ def delete_group_task(group_id, task_id):
         abort(404)
 
     if request.method == 'GET':
+        membership = session.query(GroupMember).filter(GroupMember.user_id==get_current_user().id, GroupMember.group_id==group_id).first()
+        if membership is None:
+            return {
+                'code': 'user',
+                'message': 'Unauthorized user access/No rights',
+            }, 401
         return jsonify(GroupTaskInfoSchema().dump(task))
     elif request.method == 'DELETE':
+        if get_current_user().id != group.owner_id:
+            return {
+                'code': 'user',
+                'message': 'Unauthorized user access/No rights',
+            }, 401
         session.query(GroupTask).filter(GroupTask.id==task_id).delete()
         session.commit()
-        return jsonify({'Message': 'There is no such group task now!'})
+        return jsonify({'Message': 'Deleted successfully'})
 
 
 
@@ -177,11 +262,11 @@ def delete_group_task(group_id, task_id):
 
 
 
-@group.route('/all', methods=['GET'])
-def get_all_groups():
-    session = get_session()
+# @group.route('/all', methods=['GET'])
+# def get_all_groups():
+#     session = get_session()
 
-    groups = []
-    for group in session.query(Group).all():
-        groups.append(GroupInfoSchema().dump(group))
-    return jsonify(groups)
+#     groups = []
+#     for group in session.query(Group).all():
+#         groups.append(GroupInfoSchema().dump(group))
+#     return jsonify(groups)
